Add content to Slow Cooker Mac and Cheese recipe
Added picture, ingredients, and directions of making the Slow Cooker
Mac and Cheese to the recipe page. Used headers to separate the ordered
and unordered lists used to display the ingredients and directions.
Picture is not downloaded but sourced from the website itself for
simplicity.

Before adding a link to navigate back to the mai page, I want to save
the content beforehand incase something goes wrong and I lose it all.